Switch to using docker environment for boulder server to reduce converge times
This completely switches how the boulder server is setup to using containers.
This uses the docker compose environment provided by upstream which is more
updated and doesn't require much compiling. This seems to reduce at least two
minutes off running the builds.

Since docker support on CentOS 6 is lacking, this keeps the same deployment
setup as before for that version of the platform. We still have a few hosts that
are using letsencrypt so I thought it was important to keep this support
included.

This also includes improved integration testing which deploys an apache web
server using acme certs to ensure the whole setup works properly. In addition,
it switches from modifying /etc/hosts to using a local dnsmasq server to help
with DNS resolution.